Create table address for tests
